Make playback optional
