+++ Cargo.toml
@@ -17,12 +17,16 @@ bytes = "1.4.0"
 futures-channel = "0.3.30"
 futures-util = "0.3.28"
 reqwest = { version = "0.12.5", features = ["stream", "json", "multipart"] }
-rodio = "0.17.1"
+rodio = { version = "0.17.1", optional = true }
 serde = { version = "1.0.173", features = ["derive"] }
 serde_json = "1.0.103"
 thiserror = "1.0.43"
 tokio = { version = "1.29.1", features = ["full"] }
 tokio-tungstenite = { version = "0.23.0", features = ["native-tls"] }
+
+[features]
+default = ["playback"]
+playback = ["dep:rodio"]
 
 [dev-dependencies]
 twilio = "1.1.0"
@@ -49,6 +53,3 @@ path = "examples/pronunciation_dictionaries.rs"
 [[example]]
 name = "twilio_openai"
 path = "examples/twilio_openai.rs"
-
-
-
